wip: refactored csv driver
